* Keep two decimal places for value.

+++ app/sys/common/view/chart.html.php
@@ -45,7 +45,7 @@ if($config->debug)
                     var color = nextAccentColor().toCssStr();
 
                     $row.attr('data-id', idx).find('.chart-color-dot').css('color', color);
-                    data.push({label: $row.find('.chart-label').text(), value: parseInt($row.find('.chart-value').text()), color: color, id: idx});
+                    data.push({label: $row.find('.chart-label').text(), value: parseFloat($row.find('.chart-value').text()), color: color, id: idx});
                 });
 
                 if(data.length > 1) options.scaleLabelPlacement = 'outside';
@@ -74,7 +74,7 @@ if($config->debug)
                 {
                     var $row = $(this);
                     labels.push($row.find('.chart-label').text());
-                    dataset.data.push(parseInt($row.find('.chart-value').text()));
+                    dataset.data.push(parseFloat($row.find('.chart-value').text()));
                 });
                 var data = {labels: labels, datasets: [dataset]};
                 if(labels.length) options.barValueSpacing = 5;
